+++ backend/apps/employees/models.py
@@ -8,9 +8,9 @@ class Employee(BaseModel):
 
     name = models.CharField(max_length=20, blank=False, null=False)
     last_name = models.CharField(max_length=20, blank=False, null=False)
-    identification = models.IntegerField(blank=False, null=False, unique=True)
+    identification = models.BigIntegerField(blank=False, null=False, unique=True)
     email = models.EmailField(blank=False, null=False, unique=True)
-    phone = models.IntegerField(unique=True)
+    phone = models.BigIntegerField(unique=True)
     position = models.CharField("position", max_length=20,
                                 choices=POSITIONS, default=POSITIONS.barbero)
